+++ tutorials/TRAIN_ALIGNER.py
@@ -17,7 +17,7 @@
 #
 # This notebook replicates the training setup from:
 # ```
-# cogames train -m aligner_tutorial -p tutorial
+# cogames tutorial train -m aligner_tutorial -p tutorial
 # ```
 #
 # It walks through:
@@ -491,7 +491,7 @@ else:
 # Submit the trained weights to the CoGames tournament. This uses the `tutorial` policy class
 # (same architecture as `AlignerPolicyNet`) with our saved weights.
 #
-# Prerequisites: run `cogames login` in a terminal first to authenticate.
+# Prerequisites: run `cogames auth login` in a terminal first to authenticate.
 
 # %%
 POLICY_NAME = "my-aligner"  # Change this to your desired policy name